requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.29
psycopg[binary]==3.1.18
alembic==1.13.1

# Environment & config
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1

# OpenAI
openai==2.21.0

# Utils
requests==2.31.0
httpx==0.27.0
pypdf
python-multipart

# Optional but useful
tenacity==8.2.3
# orjson==3.9.15
